requirements.txt
flask
flask-cors
werkzeug
pyjwt
feedparser
requests
sentence-transformers
scikit-learn
numpy
apscheduler
supabase
psycopg2-binary==2.9.7
python-dotenv==1.0.0